pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "geo-research-agent-247"
version = "247.0.0"
description = "Fully automatic AI research agent for Geoscientists, Remote Sensing researchers, and GIScientists"
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"
keywords = ["GIS", "geoscience", "remote-sensing", "AI", "research-agent", "GWR", "MGWR", "spatial-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.6",
    "rasterio>=1.3",
    "mgwr>=2.2.1",
    "statsmodels>=0.14",
    "numpy>=1.26",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "scikit-learn>=1.4",
    "arxiv>=2.1",
]

[project.optional-dependencies]
full = [
    "openai>=1.50.0",
    "pysal>=23.7",
    "libpysal>=4.9",
    "esda>=2.5",
    "pykrige>=1.7",
    "contextily>=1.5",
    "folium>=0.15",
    "earthpy>=0.9",
    "sentinelsat>=1.1",
    "pypandoc>=1.13",
    "jinja2>=3.1",
    "jupyter>=1.0",
]
dev = [
    "pytest>=7.4",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
geo-research = "core.orchestrator:main"
geo-benchmark = "GeoBenchmark.run_benchmark:main"
geo-download-data = "GeoBenchmark.download_data:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
